Split the common classes into separate modules
